refactoring: Network初期化コードの修正

+++ overlay_rootfs/scripts/network_init.sh
@@ -17,18 +17,23 @@ if [ -f /atom/system/driver/mmc_detect_test.ko ]; then
   VENDERID=`cat /sys/bus/mmc/devices/mmc1\:0001/mmc1\:0001\:1/vendor`
 fi
 if [ "0x024c" = "$VENDERID" ]; then
-  insmod /atom/system/driver/rtl8189ftv.ko rtw_power_mgnt=0 rtw_enusbss=0
+  insmod /atom/system/driver/rtl8189ftv.ko
 elif [ "0x007a" = "$VENDERID" ]; then
-  [ -f /atom/usr/share/atbm603x_conf/atbm_txpwer_dcxo_cfg.txt ] && cp /atom/usr/share/atbm603x_conf/atbm_txpwer_dcxo_cfg.txt /tmp
-  [ -f /atom/usr/share/atbm603x_conf/set_rate_power.txt ] && cp /atom/usr/share/atbm603x_conf/set_rate_power.txt /tmp
-  [ -f /lib/modules/atbm603x_wifi_sdio.ko ] && insmod /lib/modules/atbm603x_wifi_sdio.ko
-  sleep 1
-  if [ ! -f /sys/module/atbm603x_wifi_sdio/parameters/fw_ver ]; then
-    sync
-    echo 3 > /proc/sys/vm/drop_caches
-    [ -f /lib/modules/atbm603x_wifi_sdio.ko ] && insmod /lib/modules/atbm603x_wifi_sdio.ko
+  if [ -f /atom/therdlib/atbm603x_wifi_sdio.ko ] ; then
+    insmod /atom/therdlib/atbm603x_wifi_sdio.ko
+  elif [ -f /atom/system/driver/atbm603x_wifi_sdio.ko ] ; then
+    [ -f /atom/usr/share/atbm603x_conf/atbm_txpwer_dcxo_cfg.txt ] && cp /atom/usr/share/atbm603x_conf/atbm_txpwer_dcxo_cfg.txt /tmp
+    [ -f /atom/usr/share/atbm603x_conf/set_rate_power.txt ] && cp /atom/usr/share/atbm603x_conf/set_rate_power.txt /tmp
+    insmod /atom/system/driver/atbm603x_wifi_sdio.ko
+    if [ ! -f /sys/module/atbm603x_wifi_sdio/parameters/fw_ver ]; then
+      sync
+      echo 3 > /proc/sys/vm/drop_caches
+      insmod /atom/system/driver/atbm603x_wifi_sdio.ko
+    fi
+  else
+    insmod /lib/modules/atbm603x_wifi_sdio.ko
   fi
-  echo "LOG_ERR=OFF LOG_WARN=ON LOG_LMAC=ON LOG_SCAN=OFF" > /sys/module/atbm603x_wifi_sdio/atbmfs/atbm_printk_mask
+  echo "LOG_ERR=OFF LOG_SCAN=OFF" > /sys/module/atbm603x_wifi_sdio/atbmfs/atbm_printk_mask 2> /dev/null
 elif [ "0x5653" = "$VENDERID" ]; then
   insmod /atom/system/driver/ssv6x5x.ko stacfgpath=/atom/system/driver/ssv6x5x-wifi.cfg
 elif [ "0x424c" = "$VENDERID" ]; then
